General_info
in order to find out the number of files per decade for a corpus(e.g.
the ECCO/EEBO-TCP),  and specify if they have 'paratextual DIV's in the
front, back, both or none in order to find their number and
percentage(out of total of files), this script outputs into the
text_file, in each line, the filename, cleaned year, and state of
paratext (frontonly, backonly, both or none).